simple Modal component for responsive motionPreset
